plugins/device_tree: Refactor and use `AnalysisPluginV0`
Version bumped because unknown-model is now also part of the summary.
Defining the schema allowed us to clean up much of the logic.
While there is some arguable code left this is a step in the right
direction.
Also required a change in the architecture_detection plugin to
accomodate that the result can be None.